project/mlops_project/web_service/requirements_web_service.txt
flask
gunicorn
pandas
mlflow
python-dotenv
autogluon.tabular[all]